UBER-953: Fix related issues
