+++ plugins/lead-assets/lang/en.json
@@ -31,6 +31,7 @@
     "Members": "Members",
     "UnAssign": "Unassign",
     "ConfigLabel": "CRM",
-    "ConfigDescription": "Extension for Customer relation management"
+    "ConfigDescription": "Extension for Customer relation management",
+    "EditFunnel": "Edit Funnel"
   }
 }
