+++ plugins/tracker-assets/lang/en.json
@@ -283,7 +283,9 @@
     "IssueNotificationChangedProperty": "{senderName} changed {property} to \"{newValue}\"",
     "IssueNotificationMessage": "{senderName}: {message}",
     "PreviousAssigned": "Previously assigned",
-    "IssueAssigneedToYou": "Assigned to you"
+    "IssueAssigneedToYou": "Assigned to you",
+    "RelatedIssueTargetDescription": "Related issue project target for Class or Space",
+    "MapRelatedIssues": "Configure Related issue default projects"
   },
   "status": {}
 }
