+++ plugins/tracker-assets/lang/ru.json
@@ -283,7 +283,9 @@
     "IssueNotificationChangedProperty": "{senderName} изменил {property} на \"{newValue}\"",
     "IssueNotificationMessage": "{senderName}: {message}",
     "PreviousAssigned": "Ранее назначенные",
-    "IssueAssigneedToYou": "Назначено вам"
+    "IssueAssigneedToYou": "Назначено вам",
+    "RelatedIssueTargetDescription": "Настройка проекта по умолчанию для Класса или пространства",
+    "MapRelatedIssues": "Настроить проекты по умолчанию для связанных задач"
   },
   "status": {}
 }
