+++ plugins/tracker-resources/package.json
@@ -64,6 +64,7 @@
     "@hcengineering/workbench-resources": "^0.6.1",
     "@hcengineering/activity-resources": "^0.6.1",
     "@hcengineering/activity": "^0.6.0",
-    "@hcengineering/query": "^0.6.7"
+    "@hcengineering/query": "^0.6.7",
+    "@hcengineering/preference": "^0.6.8"
   }
 }
