+++ plugins/tracker-resources/src/plugin.ts
@@ -226,7 +226,6 @@ export default mergeIds(trackerId, tracker, {
     AddRelatedIssue: '' as IntlString,
     RelatedIssuesNotFound: '' as IntlString,
     RelatedIssue: '' as IntlString,
-    RelatedIssues: '' as IntlString,
     BlockedIssue: '' as IntlString,
     BlockingIssue: '' as IntlString,
     BlockedBySearchPlaceholder: '' as IntlString,
@@ -301,7 +300,9 @@ export default mergeIds(trackerId, tracker, {
     NoStatusFound: '' as IntlString,
     CreateMissingStatus: '' as IntlString,
     UnsetParent: '' as IntlString,
-    PreviousAssigned: '' as IntlString
+    PreviousAssigned: '' as IntlString,
+    EditRelatedTargets: '' as IntlString,
+    RelatedIssueTargetDescription: '' as IntlString
   },
   component: {
     NopeComponent: '' as AnyComponent,
